+++ packages/headless/doc-parser/src/function-resolver.ts
@@ -1,0 +1,46 @@
+import {
+  ApiEntryPoint,
+  ApiFunction,
+  ApiInterface,
+} from '@microsoft/api-extractor-model';
+import {DocComment} from '@microsoft/tsdoc';
+import {findApi} from './api-finder';
+import {ObjEntity} from './entity';
+import {buildEntity, buildFuncEntity} from './entity-builder';
+import {resolveParams} from './function-param-resolver';
+import {resolveInterfaceMembers} from './interface-resolver';
+
+export function resolveFunction(
+  entry: ApiEntryPoint,
+  fn: ApiFunction,
+  shallowParamIndices: number[]
+) {
+  const params = resolveParams(entry, fn, shallowParamIndices);
+  const returnTypeText = fn.returnTypeExcerpt.text;
+  const returnTypeInterface = findApi(entry, returnTypeText) as ApiInterface;
+
+  const returnType = buildObjEntityFromInterface(entry, returnTypeInterface);
+
+  return buildFuncEntity({
+    name: fn.name,
+    comment: (fn.tsdocComment as unknown) as DocComment,
+    params,
+    returnType,
+  });
+}
+
+function buildObjEntityFromInterface(
+  entryPoint: ApiEntryPoint,
+  apiInterface: ApiInterface
+): ObjEntity {
+  const name = apiInterface.name;
+  const members = resolveInterfaceMembers(entryPoint, apiInterface);
+  const entity = buildEntity({
+    name,
+    type: name,
+    isOptional: false,
+    comment: (apiInterface.tsdocComment as unknown) as DocComment,
+  });
+
+  return {...entity, members};
+}